split network code into more modules
