+++ src/network/mod.rs
@@ -1,0 +1,13 @@
+/// Module for server-specific network code
+pub mod server;
+
+/// Module for client-specific network code
+pub mod client;
+
+/// Module for network code common between server and client
+mod common;
+
+/// Re-export everything in common as if it was here
+pub use common::*;
+
+
